+++ saritasa_invocations/django.py
@@ -1,4 +1,5 @@
 import collections.abc
+import contextlib
 import os
 import pathlib
 
@@ -146,23 +147,56 @@ def createsuperuser(
     password: str = "",
 ) -> None:
     """Create superuser."""
-    printing.print_success("Django: Create superuser")
     config = _config.Config.from_context(context)
+    email_source = "cmd"
+    username_source = "cmd"
+    if not email:
+        with contextlib.suppress(invoke.Failure):
+            output = context.run(
+                "git config user.email",
+                echo=False,
+                hide="out",
+            )
+            if output and (email := output.stdout.replace(" ", "").strip()):
+                email_source = "git"
+            else:
+                email = config.django.default_superuser_email
+                email_source = "config"
+    if not username:
+        with contextlib.suppress(invoke.Failure):
+            output = context.run(
+                "git config user.name",
+                echo=False,
+                hide="out",
+            )
+            if output and (username := output.stdout.replace(" ", "").strip()):
+                username_source = "git"
+            else:
+                username = config.django.default_superuser_username
+                username_source = "config"
+    if not password:
+        password = config.django.default_superuser_password
+
+    printing.print_success(
+        "Django: Creating superuser with the following ->\n"
+        f"{email=} from {email_source}\n"
+        f"{username=} from {username_source}\n",
+    )
+
     responder_email = invoke.FailingResponder(
         pattern=rf"{config.django.verbose_email_name}.*: ",
-        response=(email or config.django.default_superuser_email) + "\n",
+        response=f"{email}\n",
         sentinel="That Email address is already taken.",
     )
     responder_user_name = invoke.Responder(
         pattern=rf"{config.django.verbose_username_name}.*: ",
-        response=(username or config.django.default_superuser_username) + "\n",
+        response=f"{username}\n",
     )
     password_pattern = config.django.verbose_password_name
     responder_password = invoke.Responder(
         pattern=rf"({password_pattern}: )|({password_pattern} \(again\): )",
-        response=(password or config.django.default_superuser_password) + "\n",
+        response=f"{password}\n",
     )
-
     try:
         manage(
             context,